+++ apps/backend/prisma/migrations/20250917193043_store_price_id/migration.sql
@@ -1,0 +1,5 @@
+-- AlterTable
+ALTER TABLE "OneTimePurchase" ADD COLUMN     "priceId" TEXT;
+
+-- AlterTable
+ALTER TABLE "Subscription" ADD COLUMN     "priceId" TEXT;