+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/sidebar-layout.tsx
@@ -35,6 +35,7 @@ import {
   Mail,
   Menu,
   Palette,
+  Receipt,
   Settings,
   Settings2,
   ShieldEllipsis,
@@ -246,10 +247,17 @@ const navigationItems: (Label | Item | Hidden)[] = [
     type: 'label',
   },
   {
-    name: "Offers & Items",
-    href: "/payments/offers-and-items",
-    regex: /^\/projects\/[^\/]+\/payments\/offers-and-items$/,
+    name: "Offers",
+    href: "/payments/offers",
+    regex: /^\/projects\/[^\/]+\/payments\/offers$/,
     icon: CreditCard,
+    type: 'item',
+  },
+  {
+    name: "Transactions",
+    href: "/payments/transactions",
+    regex: /^\/projects\/[^\/]+\/payments\/transactions$/,
+    icon: Receipt,
     type: 'item',
   },
   {