+++ examples/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/example-demo-app
 
+## 2.8.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.39
+  - @stackframe/stack-ui@2.8.39
+  - @stackframe/stack@2.8.39
+
 ## 2.8.38
 
 ### Patch Changes