+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stackframe/js-example
 
+## 2.8.39
+
+### Patch Changes
+
+- @stackframe/js@2.8.39
+
 ## 2.8.38
 
 ### Patch Changes