+++ packages/template/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/stack
 
+## 2.8.39
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.39
+  - @stackframe/stack-ui@2.8.39
+  - @stackframe/stack-sc@2.8.39
+
 ## 2.8.38
 
 ### Patch Changes